fixed sphere/sphere collision detection
Former-commit-id: ca298445ae136f8db70def8ad77ef5871790da49

+++ Space/ComponentSystem/Entities/EntityFactory.cs
@@ -23,7 +23,7 @@ namespace Space.ComponentSystem.Entities
 
             var collidable = new CollidableSphere();
             collidable.Radius = shipData.Radius;
-            collidable.CollisionGroup = (byte)playerNumber;
+            collidable.CollisionGroup = playerNumber;
             entity.AddComponent(collidable);
 
             var modules = new EntityModules<EntityAttributeType>();
@@ -110,7 +110,7 @@ namespace Space.ComponentSystem.Entities
             var avatar = emitter.GetComponent<Avatar>();
             if (avatar != null)
             {
-                collidable.CollisionGroup = (byte)avatar.PlayerNumber;
+                collidable.CollisionGroup = avatar.PlayerNumber;
             }
             entity.AddComponent(collidable);
